Code_Gen_Agent/requirements.txt
agno
pylance 
lancedb 
tantivy 
openai 
sqlalchemy
bs4
python-dotenv
huggingface_hub
mistralai